Added the fourth (fifth) tutorial to C++
I added the fourth proper tutorial to the C++ files and fifth one in total (including the setup. I also changed the `int three = 4` to `int three = 3` in all the files where it was written wrong by me. I also did the same for the `int four`. This was earlier requested by @BD103  to I am adding him as a co-author.

Co-Authored-By: BD103 <59022059+BD103@users.noreply.github.com>